Separate compiler definition
